requirements.txt
discord
python-a2s
python-dotenv
